add[search functionality]:allLeaves
