+++ frontend/src/style.css
@@ -6,6 +6,7 @@
   --dark: #000000;
 }
 
+
 .bg {
   background-color: #191c24;
